Move FFDW createMetadata function to shared utils package
