+++ packages/utils/package.json
@@ -8,6 +8,7 @@
   },
   "exports": {
     "./buildImageSizeProp": "./src/buildImageSizeProp.ts",
+    "./createMetaData": "./src/createMetaData.ts",
     "./cmsConfigUtils": "./src/cmsConfigUtils.ts",
     "./convertYamlToJson": "./src/convertYamlToJson.ts",
     "./dateUtils": "./src/dateUtils.ts",
